unseen notifications logic

+++ notifications-handler.php
@@ -1,6 +1,7 @@
 <?php
     require_once 'includes/db-setup.php'; // includes config.php where $config is defined (and so, the base_url)
     require_once 'includes/config-session.php'; // includes config file as well
+    require_once 'login/model.php';
 
     $pdo = connect_db();
 
@@ -15,10 +16,13 @@
     if (isset($_POST["mark_seen"])) {
         $id = $_POST["mark_seen"];
 
-        $query = "UPDATE notifications SET seen = 1 WHERE id = :id";
+        $query = "UPDATE notifications SET seen = 1 WHERE id = :id AND user_id = :user_id";
         $stmt = $pdo->prepare($query);
         $stmt->bindParam(":id", $id);
+        $stmt->bindParam(":user_id", $_SESSION["user_id"]);
         $stmt->execute();
+
+        $_SESSION["has_unseen_notifications"] = has_unseen_notifications($pdo, $_SESSION["user_id"]);
 
         echo json_encode(array("success" => true));
         die();
@@ -27,33 +31,32 @@
 
     // Delete all notifications
     if (isset($_POST["delete_all"])) {
-        $query = "DELETE FROM notifications";
+        $query = "DELETE FROM notifications WHERE user_id = :user_id";
         $stmt = $pdo->prepare($query);
+        $stmt->bindParam(":user_id", $_SESSION["user_id"]);
         $stmt->execute();
+
+        $_SESSION["has_unseen_notifications"] = false;
 
         echo json_encode(array("success" => true));
         die();
     }
+    // =========================================
 
     // Fetch notifications
     if (!isset($_SESSION["user_id"])) {
-        echo json_encode(array("success" => false));
         header("Location: " . $config["base_url"]);
         die();
     }
 
-    $is_admin = $_SESSION["user_role"] == "admin";
-
-    $query = "SELECT * FROM notifications";
-    if (!$is_admin) { $query .= " WHERE user_id = :user_id"; }
-
+    $query = "SELECT * FROM notifications WHERE user_id = :user_id";
     $stmt = $pdo->prepare($query);
-    if (!$is_admin) { $stmt->bindParam(":user_id", $_SESSION["user_id"]);}
-
+    $stmt->bindParam(":user_id", $_SESSION["user_id"]);
     $stmt->execute();
+
     $notifications_array = $stmt->fetchAll();
     
     $arr = array("notifications" => array_reverse($notifications_array));
     echo json_encode($arr);
     die();
-    // =========================================
+    // ========================================
